Removed possibility of double entry in hosts file

+++ selfcontrol.py
@@ -5,23 +5,33 @@ import sys
 import subprocess
 import time
 
-def startSelfControl(blockedListFilePath,hostsFilePath):
-    '''Adds website lines from blocked.txt to /etc/hosts'''
+def addLinesToFile(blockedListFilePath,hostsFilePath):
+    '''Adds website lines from blocked.txt to /etc/hosts if not already present'''
 
     blockedListFile = open(blockedListFilePath,"r")
     hostsFile = open(hostsFilePath,"a")
     hostsFile.write("\n#SELFCONTROL BLOCK START\n")
     for site in blockedListFile:
         addString = site
-        if site[0:3] != "www":
+        if site[0:4] != "www.":
             addString = "www."+addString
         hostsFile.write("0.0.0.0 " +addString)
         hostsFile.write("::0 "+addString)
     hostsFile.write("#SELFCONTROL BLOCK END\n\n")
 
+
+def linesAlreadyPresent(hostsFilePath):
+    '''Checks if lines are already present in hosts file'''
+    hostsFile = open(hostsFilePath,"r")
+
+    for i in hostsFile:
+        if i.rstrip("\n") == "#SELFCONTROL BLOCK START":
+            return True
+    return False
+
+
 def waitForCompletion(hh,mm):
     '''Runs a countdown timer'''
-
     startTime = time.time()
     currTime = time.time()
     reqTime = hh*3600+mm*60
@@ -54,9 +64,10 @@ def main():
     hostsFilePath = "/etc/hosts"
     os.system("cp "+hostsFilePath+" hosts_backup")
     os.system("cp hosts_backup hosts_new")
-    startSelfControl(blockedListFilePath,"hosts_new")
-    os.system("cp hosts_new "+hostsFilePath)
-    os.system("rm hosts_new")
+    if not linesAlreadyPresent(hostsFilePath):
+        addLinesToFile(blockedListFilePath,"hosts_new")
+        os.system("cp hosts_new "+hostsFilePath)
+        os.system("rm hosts_new")
     waitForCompletion(hh,mm)
     endSelfControl(hostsFilePath)
